REFACTOR: update home page, changed name of website
