+++ frontend/learning-rate/src/app/page.tsx
@@ -1,16 +1,12 @@
 'use client'
 import Hero from "@/app/components/hero";
-import PGInfo from "@/app/components/playground_info";
-import LessonsInfo from "@/app/components/lessons_info";
-import ChatComponent from "@/app/components/chat_component";
+import ChatbotInfo from "@/app/components/chatbot_info";
 
 export default function Home() {
   return (
     <main>
       <Hero />
-      <PGInfo />
-      <ChatComponent />
-      <LessonsInfo />
+      <ChatbotInfo />
     </main>
   );
 };
